backend-flask/requirements.txt
flask
gunicorn
langchain
langchain-community
langchain-core
langsmith
langchain-openai
openai
PyPDF2
python-dotenv
faiss-cpu
tiktoken
werkzeug
flask-cors